% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{all_ages}
\alias{all_ages}
\title{All Installed Package Ages}
\usage{
all_ages(units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  decreasing = FALSE, ...)
}
\arguments{
\item{units}{character string specifying the units (e.g., \code{"days"}). Can
be abbreviated.}

\item{decreasing}{Logical indicating whether or not to list the ages
in decreasing order (i.e., youngest to oldest). Default is \code{FALSE}.}

\item{...}{Additional optional arguments to be passed on to
\code{pkg.age}.}
}
\value{
A data frame with two columns: \code{package} and \code{age}.
}
\description{
List the age of all your installed packages.
}
\examples{
\dontrun{
all_ages(units = "weeks", decreasing = TRUE)
}
}
