% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{all_bdays}
\alias{all_bdays}
\title{All Installed Package Birthdays}
\usage{
all_bdays(decreasing = FALSE, ...)
}
\arguments{
\item{decreasing}{Logical indicating whether or not to list the birthdays
in decreasing order (i.e., youngest to oldest). Default is \code{FALSE}.}

\item{...}{Additional optional arguments to be passed on to
\code{pkg_bday}.}
}
\value{
A data frame with two columns: \code{package} and \code{birthday}.
}
\description{
List the birthdays of all your installed packages.
}
\examples{
\dontrun{
all_bdays()
}
}
