% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{pkg_age}
\alias{pkg_age}
\title{Determine the Age of An Installed Package}
\usage{
pkg_age(package, units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  ...)
}
\arguments{
\item{package}{Character string specifying the name of the package.}

\item{units}{character string specifying the units (e.g., \code{"days"}). Can
be abbreviated.}

\item{...}{Additional optional arguments to be passed on to
\code{pkg_bday}.}
}
\value{
A \code{"difftime"} object representing the age of the package. Can
be coerced to a double using \code{as.double}.
}
\description{
Determine the age of a particular installed package. The age is simply the
time difference (using \code{difftime}) between today's date and the
package's very first appearance on CRAN.
}
\examples{
\dontrun{
pkg_age("rpart")
cat("rpart is", as.double(pkg.age("rpart")) / 365, "years old today!\\n")
}
}
