% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{pkg_bday}
\alias{pkg_bday}
\title{Determine an Installed Packages Birthday}
\usage{
pkg_bday(package, ...)
}
\arguments{
\item{package}{Character string specifying the name of the package.}

\item{...}{Additional optional arguments. (Currently ignored.)}
}
\value{
A character string representing the birthday (i.e.,
\code{"mm/dd/yyyy"}) of the package.
}
\description{
Determine the birthday of a particular installed package. The birthdate is
simply the package's very first appearance on CRAN.
}
\examples{
pkg_bday("rpart")
}
