% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{upcoming_bdays}
\alias{upcoming_bdays}
\title{Packages with Upcoming Birthdays}
\usage{
upcoming_bdays(days = 30)
}
\arguments{
\item{days}{Integer specifying the number of days out from the current date
to consider.}
}
\value{
A data frame with two columns: \code{package} and \code{birthday}
}
\description{
Determine which of your installed packages have an upcoming birthday.
}
\note{
Based on a modified version of the internal
\code{\link[devtools]{devtools}} function \code{package_find_repo}.
}
\examples{
\dontrun{
upcoming_bdays()
}
}
